Extract fake platform impl to a separate file
